# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meals:
    post:
      operationId: createMeal
      tags: [meals]
      summary: Save a meal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMealRequest"
      responses:
        "201":
          description: Meal saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealEntry"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /meals/today:
    get:
      operationId: getTodayMeals
      tags: [meals]
      summary: Get today's meal entries
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Today's meals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodayMealsResponse"
  /meals/history:
    get:
      operationId: getMealHistory
      tags: [meals]
      summary: Get meal history (past days)
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Meal history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealEntry"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CreateMealRequest:
      type: object
      properties:
        userId:
          type: string
          description: User UUID v4
        photoUrl:
          type: string
          nullable: true
          description: publicUrl from the upload-url response, or null if no photo
        aiResponse:
          type: object
          nullable: true
          description: Raw AI analysis payload for audit purposes
        userCorrections:
          type: object
          nullable: true
          description: User-edited ingredient portions (if any)
        totalCalories:
          type: integer
        totalProteinG:
          type: number
        totalCarbsG:
          type: number
        totalFatG:
          type: number
        mealName:
          type: string
        localHour:
          type: integer
          minimum: 0
          maximum: 23
          description: |
            Client's current local hour (0–23). Used to assign mealType in user's timezone.
            Falls back to server UTC hour if omitted.
      required:
        - userId
        - totalCalories
        - totalProteinG
        - totalCarbsG
        - totalFatG
        - mealName
    MealEntry:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        photoUrl:
          type: string
          nullable: true
        aiResponse:
          type: object
          nullable: true
        userCorrections:
          type: object
          nullable: true
        totalCalories:
          type: integer
        totalProteinG:
          type: number
        totalCarbsG:
          type: number
        totalFatG:
          type: number
        mealName:
          type: string
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - totalCalories
        - totalProteinG
        - totalCarbsG
        - totalFatG
        - mealName
        - mealType
        - createdAt
    TodayMealsResponse:
      type: object
      properties:
        meals:
          type: array
          items:
            $ref: "#/components/schemas/MealEntry"
        totalCalories:
          type: integer
        totalProteinG:
          type: number
        totalCarbsG:
          type: number
        totalFatG:
          type: number
        streak:
          type: integer
      required:
        - meals
        - totalCalories
        - totalProteinG
        - totalCarbsG
        - totalFatG
        - streak
...